requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
fastapi[all]==0.112.0
pydantic-settings==2.4.0
python-telegram-bot==21.4
requests==2.32.3
sentry-sdk[fastapi]==2.12.0
SQLAlchemy[asyncio]==2.0.31
